Added Delete feature on the application

+++ database/migrations/2026_03_17_100434_add_title_to_books_table.php
@@ -8,14 +8,18 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('books', function (Blueprint $table) {
-            $table->dropColumn('title');
+            if (!Schema::hasColumn('books', 'title')) {
+                $table->string('title')->after('user_id');
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('books', function (Blueprint $table) {
-            $table->string('title')->nullable();
+            if (Schema::hasColumn('books', 'title')) {
+                $table->dropColumn('title');
+            }
         });
     }
 };
